internal changes required to improve handling time indexing in data files

+++ src/datasources/ascii/asciidatainterfaces.h
@@ -47,7 +47,7 @@ public:
   bool isValid(const QString& field) const { return ascii._fieldLookup.contains( field ); }
 
   // T specific
-  const DataVector::DataInfo dataInfo(const QString&, int frame = 0) const;
+  const DataVector::DataInfo dataInfo(const QString&, double frame = 0) const;
   void setDataInfo(const QString&, const DataVector::DataInfo&) {}
 
   // meta data
@@ -60,7 +60,7 @@ public:
 
 
 //-------------------------------------------------------------------------------------------
-const DataVector::DataInfo DataInterfaceAsciiVector::dataInfo(const QString &field, int frame) const
+const DataVector::DataInfo DataInterfaceAsciiVector::dataInfo(const QString &field, double frame) const
 {
   Q_UNUSED(frame)
   if (!ascii._fieldLookup.contains(field))
@@ -126,7 +126,7 @@ public:
   bool isValid(const QString&) const;
 
   // T specific
-  const DataString::DataInfo dataInfo(const QString&, int frame=0) const { Q_UNUSED(frame) return DataString::DataInfo(); }
+  const DataString::DataInfo dataInfo(const QString&, double frame=0) const { Q_UNUSED(frame) return DataString::DataInfo(); }
   void setDataInfo(const QString&, const DataString::DataInfo&) {}
 
   // meta data
