+++ src/datasources/bis/bisdatasource.cpp
@@ -65,7 +65,7 @@ public:
   bool isValid(const QString&) const;
 
   // T specific
-  const DataMatrix::DataInfo dataInfo(const QString&, int frame=0) const;
+  const DataMatrix::DataInfo dataInfo(const QString&, double frame=0) const;
   void setDataInfo(const QString&, const DataMatrix::DataInfo&) {}
 
   // meta data
@@ -90,7 +90,7 @@ void DataInterfaceBISMatrix::init()
 
 
 
-const DataMatrix::DataInfo DataInterfaceBISMatrix::dataInfo(const QString& matrix, int frame) const
+const DataMatrix::DataInfo DataInterfaceBISMatrix::dataInfo(const QString& matrix, double frame) const
 {
   if (_bis._bisfile->status != BIS_OK) {
     return DataMatrix::DataInfo();
@@ -188,7 +188,7 @@ public:
   bool isValid(const QString&) const;
 
   // T specific
-  const DataVector::DataInfo dataInfo(const QString&, int frame=0) const;
+  const DataVector::DataInfo dataInfo(const QString&, double frame=0) const;
   void setDataInfo(const QString&, const DataVector::DataInfo&) {}
 
   // meta data
@@ -212,7 +212,7 @@ void DataInterfaceBISVector::init()
 
 
 
-const DataVector::DataInfo DataInterfaceBISVector::dataInfo(const QString& vector, int) const
+const DataVector::DataInfo DataInterfaceBISVector::dataInfo(const QString& vector, double) const
 {
   if (_bis._bisfile->status != BIS_OK) {
     return DataVector::DataInfo();
@@ -229,9 +229,9 @@ const DataVector::DataInfo DataInterfaceBISVector::dataInfo(const QString& vecto
 int DataInterfaceBISVector::read(const QString& field, DataVector::ReadInfo& p)
 {
 
-  int f0 = p.startingFrame;
-  int nr = p.numberOfFrames;
-  int nf = _bis._nframes;
+  qint64 f0 = (qint64)p.startingFrame;
+  qint64 nr = (qint64)p.numberOfFrames;
+  qint64 nf = _bis._nframes;
 
   if (f0>nf) {
     return 0;
@@ -246,10 +246,10 @@ int DataInterfaceBISVector::read(const QString& field, DataVector::ReadInfo& p)
   }
 
   if ( field=="INDEX" ) {
-    for (int i=0; i<nr; ++i ) {
+    for (qint64 i=0; i<nr; ++i ) {
       p.data[i] = i + f0;
     }
-    return nr;
+    return (int)nr;
   }
 
   return 0;
