+++ src/datasources/fitstable/fitstable.h
@@ -42,7 +42,7 @@ class FitsTableSource : public Kst::DataSource {
 
     int readString(QString *stringValue, const QString& stringName);
 
-    int readField(double *v, const QString& field, int s, int n);
+    int readField(double *v, const QString& field, double s, double n);
 
     int readMatrix(double *v, const QString& field);
